+++ src/main/java/fr/insee/genesis/controller/rest/responses/RawResponseController.java
@@ -30,6 +30,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.time.Instant;
@@ -45,6 +47,8 @@ public class RawResponseController {
 
     private static final String SUCCESS_MESSAGE = "Interrogation %s saved";
     private static final String INTERROGATION_ID = "interrogationId";
+    public static final String NB_DOCS_WITH_FORMATTED = "%d document(s) processed, including %d FORMATTED after data verification for collectionInstrumentId %s";
+    public static final String NB_DOCS = "%d document(s) processed for collectionInstrumentId %s";
     private final LunaticJsonRawDataApiPort lunaticJsonRawDataApiPort;
     private final RawResponseApiPort rawResponseApiPort;
     private final RawResponseInputRepository rawRepository;
@@ -115,9 +119,9 @@ public class RawResponseController {
         try {
             DataProcessResult result = rawResponseApiPort.processRawResponses(collectionInstrumentId, interrogationIdList, errors);
             return result.formattedDataCount() == 0 ?
-                    ResponseEntity.ok("%d document(s) processed".formatted(result.dataCount()))
-                    : ResponseEntity.ok("%d document(s) processed, including %d FORMATTED after data verification"
-                    .formatted(result.dataCount(), result.formattedDataCount()));
+                    ResponseEntity.ok(NB_DOCS.formatted(result.dataCount(), collectionInstrumentId))
+                    : ResponseEntity.ok(NB_DOCS_WITH_FORMATTED
+                    .formatted(result.dataCount(), result.formattedDataCount(), collectionInstrumentId));
         } catch (GenesisException e) {
             return ResponseEntity.status(e.getStatus()).body(e.getMessage());
         }
@@ -137,9 +141,9 @@ public class RawResponseController {
         try {
             DataProcessResult result = rawResponseApiPort.processRawResponses(collectionInstrumentId);
             return result.formattedDataCount() == 0 ?
-                    ResponseEntity.ok("%d document(s) processed".formatted(result.dataCount()))
-                    : ResponseEntity.ok("%d document(s) processed, including %d FORMATTED after data verification"
-                    .formatted(result.dataCount(), result.formattedDataCount()));
+                    ResponseEntity.ok(NB_DOCS.formatted(result.dataCount(), collectionInstrumentId))
+                    : ResponseEntity.ok(NB_DOCS_WITH_FORMATTED
+                    .formatted(result.dataCount(), result.formattedDataCount(), collectionInstrumentId));
         } catch (GenesisException e) {
             return ResponseEntity.status(e.getStatus()).body(e.getMessage());
         }
@@ -249,7 +253,7 @@ public class RawResponseController {
         log.info("Try to read raw JSONs for campaign {}, with startDate={} and endDate={} - page={} - size={}", campaignId, startDate, endDate,page,size);
         Pageable pageable = PageRequest.of(page, size);
         Page<LunaticJsonRawDataModel> rawResponses = lunaticJsonRawDataApiPort.findRawDataByCampaignIdAndDate(campaignId, startDate, endDate, pageable);
-        log.info("rawResponses={}", rawResponses.getContent().size());
+        log.info("rawResponses, lunatic-json for campaign {}, with startDate={} and endDate={} ={}", campaignId, startDate, endDate,rawResponses.getContent().size());
         return ResponseEntity.status(HttpStatus.OK).body(new PagedModel<>(rawResponses));
     }
 
@@ -264,8 +268,18 @@ public class RawResponseController {
         log.info("Try to read raw lunatic JSONs for questionnaire {} - page={} - size={}", questionnaireId, page, size);
         Pageable pageable = PageRequest.of(page, size);
         Page<LunaticJsonRawDataModel> rawResponses = lunaticJsonRawDataApiPort.findRawDataByQuestionnaireId(questionnaireId, pageable);
-        log.info("rawResponses={}", rawResponses.getContent().size());
+        log.info("rawResponses for questionnaire {} = {}",questionnaireId, rawResponses.getContent().size());
         return ResponseEntity.status(HttpStatus.OK).body(new PagedModel<>(rawResponses));
+    }
+
+    @Operation(summary = "Check existence of an interrogation")
+    @RequestMapping(value = "/responses/raw/lunatic-json/{interrogationId}", method = RequestMethod.HEAD)
+    @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<Void> existsLunaticJsonByInterrogationId(@PathVariable String interrogationId) {
+        if (lunaticJsonRawDataApiPort.existsByInterrogationId(interrogationId)) {
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @Operation(summary = "Get rawResponse JSON data from one campaign in Genesis Database, filtered by start and end dates")
@@ -281,7 +295,7 @@ public class RawResponseController {
         log.info("Try to read raw lunatic JSONs for campaign {}, with startDate={} and endDate={} - page={} - size={}", campaignId, startDate, endDate,page,size);
         Pageable pageable = PageRequest.of(page, size);
         Page<RawResponseModel> rawResponses = rawResponseApiPort.findRawResponseDataByCampaignIdAndDate(campaignId, startDate, endDate, pageable);
-        log.info("rawResponses={}", rawResponses.getContent().size());
+        log.info("rawResponses for campaign {}, with startDate={} and endDate={} ={}",campaignId, startDate, endDate, rawResponses.getContent().size());
         return ResponseEntity.status(HttpStatus.OK).body(new PagedModel<>(rawResponses));
     }
 
@@ -298,5 +312,15 @@ public class RawResponseController {
         Page<RawResponseModel> rawResponses = rawResponseApiPort.findRawResponseDataByCollectionInstrumentId(collectionInstrumentId, pageable);
         log.info("rawResponses={}", rawResponses.getContent().size());
         return ResponseEntity.status(HttpStatus.OK).body(new PagedModel<>(rawResponses));
+    }
+
+    @Operation(summary = "Check existence of an interrogation")
+    @RequestMapping(value = "/raw-responses/{interrogationId}", method = RequestMethod.HEAD)
+    @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<Void> exists(@PathVariable String interrogationId) {
+        if (rawResponseApiPort.existsByInterrogationId(interrogationId)) {
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }