+++ src/main/java/fr/insee/genesis/domain/ports/api/SurveyUnitApiPort.java
@@ -107,4 +107,6 @@ public interface SurveyUnitApiPort {
     Set<String> findCampaignIdsFrom(SurveyUnitInputDto dto);
 
     long countResponsesByQuestionnaireId(String questionnaireId);
+
+    long countDistinctInterrogationIdsByQuestionnaireAndCollectionInstrumentId(String id);
 }