+++ src/main/java/fr/insee/genesis/domain/service/rawdata/RawResponseService.java
@@ -347,6 +347,11 @@ public class  RawResponseService implements RawResponseApiPort {
         }
     }
 
+    @Override
+    public boolean existsByInterrogationId(String interrogationId) {
+        return rawResponsePersistencePort.existsByInterrogationId(interrogationId);
+    }
+
     private Map<String, Set<String>> getProcessedIdsMap(List<SurveyUnitModel> surveyUnitModels) {
         Map<String, Set<String>> processedInterrogationIdsPerQuestionnaire = new HashMap<>();
         surveyUnitModels.forEach(model ->
@@ -537,6 +542,11 @@ public class  RawResponseService implements RawResponseApiPort {
     }
 
     @Override
+    public long countDistinctInterrogationIdsByCollectionInstrumentId(String collectionInstrumentId) {
+        return rawResponsePersistencePort.countByCollectionInstrumentId(collectionInstrumentId);
+    }
+
+    @Override
     public long countByCollectionInstrumentId(String collectionInstrumentId) {
         return rawResponsePersistencePort.countByCollectionInstrumentId(collectionInstrumentId);
     }
@@ -550,6 +560,8 @@ public class  RawResponseService implements RawResponseApiPort {
     public Page<RawResponseModel> findRawResponseDataByCollectionInstrumentId(String collectionInstrumentId, Pageable pageable) {
         return rawResponsePersistencePort.findByCollectionInstrumentId(collectionInstrumentId, pageable);
     }
+
+
 
     @SuppressWarnings("unchecked")
     static void handlePairwiseCollectedVariable(